requirements.txt
streamlit
pandas
geopandas
requests
leafmap
folium
mapclassify
PyYAML
altair
